Auto post request fix, messaging from/to are links

+++ flamingo_proj/static/messaging/js/messagingKO.js
@@ -1,7 +1,9 @@
 function Message(data) {
     this.id = ko.observable(data.id);
     this.sender = ko.observable(data.sender_name);
+    this.sender_id = ko.observable(data.sender);
     this.recipient = ko.observable(data.recipient_name);
+    this.recipient_id = ko.observable(data.recipient);
     this.message_body = ko.observable(data.message_body);
     this.sent_at = ko.observable(data.sent_at);
 }
@@ -25,16 +27,21 @@ function MessagesViewModel() {
 
     self.goToMail = function(mail) {
         self.chosenFolderData(null);
-        $.get("/api/messaging/" + mail.id() + "/", {}, self.chosenMailData);
+        $.getJSON("/api/messaging/" + mail.id() + "/", {}, function(item) {
+            self.chosenMailData(new Message(item));
+        });
     }
 
     self.moveToTrash = function(mail) {
-        $.ajax({
-            url: "/api/messaging/" + mail.id() + "/",
-            type: "delete"
-        }).done(function(result){
-            self.messages.remove(mail);
-        });
+        var conf = confirm("Are you sure you want to delete this message?");
+        if(conf == true) {
+            $.ajax({
+                url: "/api/messaging/" + mail.id() + "/",
+                type: "delete"
+            }).done(function(result){
+                self.messages.remove(mail);
+            });
+        }
     }
 
     self.goToFolder('Inbox');
